refact: 計算式をその他マスタに移動しボタン構築UIへ刷新
